size_t instead of int.

+++ util.h
@@ -8,7 +8,7 @@ char *mstrcpy(char *dest, char *src);
 char *sstrcpy(char *dest, char *src);
 int is_number(const char *string);
 loff_t file_size (int fd);
-char *cmalloc(int size);
+char *cmalloc(size_t size);
 int write_from(int fp, char *buf, int size);
 void merror(char *error);
 void read_into(int fd, int block_id, char *block, int block_size);
